mock-node: use the client's protocol version in the handshake (#12870)
When using mock-node to test older versions of neard, they might not be
able to connect to each other if the protocol versions are different.
But they probably are actually compatible since the mock-node program is
doing nothing but providing serialized blocks and chunks. So just
respond with whatever protocol version the client specified in its
handshake, and add a --handshake-protocol-version flag to set it
explicitly if desired.