replace interval_duration_micros by requests_per_second as an argument to benchmark_native_transfers (#12839)
Issue [12835](https://github.com/near/nearcore/issues/12835). 

Using "requests-per-second" instead of suggested
"transactions_per_second" as this value can be used to limit the rate of
the other requests to the network as well (client data sync for
example).
Renamed the corresponding parameter in the `bencmark_mpc_sign` as well.