feat: global contract support for accounts (#12877)
This PR introduces a new `AccountContract` field in the `AccountV2`
struct, replacing `code_hash` and being defined as follows:
```
pub enum AccountContract {
    None,
    Local(CryptoHash),
    Global(CryptoHash),
    GlobalByAccount(AccountId),
}
```
It is being used to differentiate between the types of contracts stored
for the `Account`, if any.

To maintain backward compatibility, two new optional fields for global
contracts are also added in the `SerdeAccount` struct.

Part of [#12716](https://github.com/near/nearcore/issues/12716)