state-snapshot: fix protocol version calculation when creating snapshot (#12907)
We want the protocol version of the block after the head block, because
if we check the current head's protocol version, then we'll take a
snapshot the old way when we see the last block in the epoch before we
switch to the current epoch state sync. This did not cause any tests to
fail because by itself it is not strictly a bug (although suboptimal)
since we'll take another snapshot later when we see the new sync hash in
the next epoch. But it triggers a long-standing bug when multiple state
snapshot requests are made within a short time, where the state snapshot
actor might unlock flat storage updates after the first one is created
even though there's another one still to go.