fix: epoch manager deadlock (#12913)
The fact that `EpochManagerHandle` is just a wrapper over
`Arc<RwLock<EpochManager>>` became a footgun. It is tempting to call its
own methods, like we did in `get_epoch_config` and `get_shard_config`
for a long time. But this leads to grabbing lock twice in read mode.
That's obvious thing to fix but I'll also tell how I discovered it.

We didn't hit it before, as RwLock allows multiple readers. However, the
deadlock may occur if we grab write lock exactly between 2 read calls.
This is an example from official docs:
https://doc.rust-lang.org/std/sync/struct.RwLock.html

When I reduced block time to 0.6s and further, it did happen to me -
between subsequent calls inside `get_epoch_config`,
`add_validator_proposals` was also called.