fix(state-snapshots): fix skipped state snapshot after header sync (#12920)
We decide whether to take a snapshot based on the return value of
is_sync_prev_hash(), which looks at the StateSyncNewChunks column to
tell whether the given hash will be the prev_hash of the sync_hash. But
the current logic assumes that this will be called only before the sync
hash is actually found, because we clear that column after it is found
since the contents are not needed anymore.

This means that if we processed the sync header during header sync
instead of during normal block processing, then is_sync_prev_hash() will
not return true when it should. We can fix it by just checking if the
sync hash is already set before looking at the values in
StateSyncNewChunks.