feat(benchmarking): tooling for multi shard benchmarks (#12918)
This PR adds the script I've been using to benchmark `neard` TPS on a
single node (for the moment).

Benchmarks use the `synth-bm` tool under the hood.

The aim is to have an simple and reproducible way to run tests. In the
future, the structure of the script may evolve drastically, because we
want to support multi node setups like forknet.

- **simple**: just take a normal node and run a bunch of commands
```sh
export CASE=cases/local/1_node_5_shard
./bench.sh stop-nodes
./bench.sh reset
./bench.sh init
./bench.sh start-nodes
./bench.sh create-accounts
./bench.sh native-transfers
```
- **reproducible**: all configs, variables, test params are store in the
scenarios inside `case/local/`

I believe this is the best place to add this tooling in `nearcore` , but
let me know if you think otherwise.


## Update

- I have improved the way config overrides work, to reduce the amount of
repetition
- Added a way to configure log levels
- Support for localnet with N validators and 1 RPC node
- Better `README`