refactor: hold parts lock when creating new entry (#13060)
Currently we release lock right after
`maybe_insert_new_entry_in_parts_cache` and then immediately acquire it
again immediately. This can result in race condition when newly inserted
entry is evicted in between. This is highly unlikely and also won't
cause any real issues since we have `else { return Ok(()); }`, so this
fix is more of a nice-to-have refactor.